MSAPI-101 | Implement Signers options
